LayoutTests/media/media-source/media-managedmse-bufferedchange.html

<!DOCTYPE html> <!-- webkit-test-runner [ ManagedMediaSourceEnabled=true MediaSourceEnabled=true ] -->
<html>
<head>
    <title>managedmediasource</title>
    <script src="../../media/media-source/media-source-loader.js"></script>
    <script src="../../media/video-test.js"></script>
    <script>
    var loader;
    var source;
    var sourceBuffer;
    var bufferedChangeEvent;

    function loaderPromise(loader) {
        return new Promise((resolve, reject) => {
            loader.onload = resolve;
            loader.onerror = reject;
        });
    }

    window.addEventListener('load', async event => {
        try {
            findMediaElement();

            let manifests = [ 'content/test-opus-manifest.json', 'content/test-vorbis-manifest.json', 'content/test-48khz-manifest.json', 'content/test-xhe-aac-manifest.json' ];
            for (const manifest of manifests) {
                loader = new MediaSourceLoader(manifest);
                await loaderPromise(loader);
                if (ManagedMediaSource.isTypeSupported(loader.type()))
                    break;
            }

            video.disableRemotePlayback = true;
            source = new ManagedMediaSource();
            run('video.src = URL.createObjectURL(source)');
            await waitFor(source, 'sourceopen');
            waitFor(video, 'error').then(failTest);

            run('sourceBuffer = source.addSourceBuffer(loader.type())');
            testExpected('sourceBuffer.__proto__', ManagedSourceBuffer.prototype);
            testExpected('sourceBuffer.onbufferedchange !== undefined', true);

            sourceBuffer.onbufferedchange = (e) => {
                consoleWrite('onbufferedchange called.')
                bufferedChangeEvent = e;
                testExpected('bufferedChangeEvent.__proto__', BufferedChangeEvent.prototype);
            };

            run('sourceBuffer.appendBuffer(loader.initSegment())');
            await waitFor(sourceBuffer, 'update');

            consoleWrite('Append a media segment.')
            run('sourceBuffer.appendBuffer(loader.mediaSegment(0))');
            await Promise.all([waitFor(sourceBuffer, 'bufferedchange'), waitFor(sourceBuffer, 'update')]);
            testExpected('sourceBuffer.buffered.length', '1');

            consoleWrite('Clean sourcebuffer of all content.');
            run('sourceBuffer.remove(0, sourceBuffer.buffered.end(0))');
            await Promise.all([waitFor(sourceBuffer, 'bufferedchange'), waitFor(sourceBuffer, 'update')]);
            testExpected('sourceBuffer.buffered.length', '0');

            endTest();
        } catch (e) {
            failTest(`Caught exception: "${e}"`);
        }
    });
    </script>
</head>
<body>
    <video controls></video>
</body>
</html>
