LayoutTests/media/media-source/media-managedmse-bufferedchange-expected.txt


RUN(video.src = URL.createObjectURL(source))
EVENT(sourceopen)
RUN(sourceBuffer = source.addSourceBuffer(loader.type()))
EXPECTED (sourceBuffer.__proto__ == '[object ManagedSourceBuffer]') OK
EXPECTED (sourceBuffer.onbufferedchange !== undefined == 'true') OK
RUN(sourceBuffer.appendBuffer(loader.initSegment()))
EVENT(update)
Append a media segment.
RUN(sourceBuffer.appendBuffer(loader.mediaSegment(0)))
onbufferedchange called.
EXPECTED (bufferedChangeEvent.__proto__ == '[object BufferedChangeEvent]') OK
EVENT(bufferedchange)
EVENT(update)
EXPECTED (sourceBuffer.buffered.length == '1') OK
Clean sourcebuffer of all content.
RUN(sourceBuffer.remove(0, sourceBuffer.buffered.end(0)))
onbufferedchange called.
EXPECTED (bufferedChangeEvent.__proto__ == '[object BufferedChangeEvent]') OK
EVENT(bufferedchange)
EVENT(update)
EXPECTED (sourceBuffer.buffered.length == '0') OK
END OF TEST
